+++ pyprom/tests/containers/testSummitsContainer.py
@@ -10,7 +10,7 @@ from pyprom.lib.containers.summits import SummitsContainer
 from pyprom.lib.locations.saddle import Saddle
 from pyprom.lib.locations.summit import Summit
 from pyprom.tests.getData import gettestzip
-from pyprom.domain import Domain
+from pyprom.domain_map import DomainMap
 from pyprom.dataload import GDALLoader
 
 
@@ -122,7 +122,7 @@ class SummitsContainerTests(unittest.TestCase):
         datafile = GDALLoader('/tmp/N44W072.hgt')
         datamap = datafile.datamap
         someslice = datamap.subset(0, 0, 30, 30)
-        domain = Domain(someslice)
+        domain = DomainMap(someslice)
         domain.run()
         summits = domain.summits
         summit = summits[3]
@@ -133,7 +133,7 @@ class SummitsContainerTests(unittest.TestCase):
         self.assertEqual(newSummit.latitude, summit.latitude)
         self.assertEqual(newSummit.longitude, summit.longitude)
         self.assertEqual(newSummit.elevation, summit.elevation)
-        self.assertEqual(newSummit.multiPoint, summit.multiPoint)
+        self.assertEqual(newSummit.multipoint, summit.multipoint)
         self.assertEqual(newSummit.edgeEffect, summit.edgeEffect)
         self.assertEqual(newSummit.edgePoints, summit.edgePoints)
         self.assertEqual(newSummit.id, summit.id)
@@ -143,7 +143,7 @@ class SummitsContainerTests(unittest.TestCase):
         Ensure multipoint() returns all multipoint Summits
         """
         s1 = Summit(1, 1, 1)
-        s1.multiPoint = ["bogus_but_ok_for_test"]
+        s1.multipoint = ["bogus_but_ok_for_test"]
         s2 = Summit(2, 2, 2)
         container = SummitsContainer([s1, s2])
         self.assertEqual(container.multipoints, [s1])