+++ pyprom/tests/locations/testSaddle.py
@@ -7,7 +7,7 @@ the LICENSE file that accompanies it.
 
 import unittest
 
-from pyprom.domain import Domain
+from pyprom.domain_map import DomainMap
 from pyprom.lib.locations.saddle import Saddle
 from pyprom.lib.locations.summit import Summit
 from pyprom.lib.containers.linker import Linker
@@ -53,7 +53,7 @@ class SaddleTests(unittest.TestCase):
         self.assertEqual(newSaddle.latitude, saddle.latitude)
         self.assertEqual(newSaddle.longitude, saddle.longitude)
         self.assertEqual(newSaddle.elevation, saddle.elevation)
-        self.assertEqual(newSaddle.multiPoint, saddle.multiPoint)
+        self.assertEqual(newSaddle.multipoint, saddle.multipoint)
         self.assertEqual(newSaddle.highShores, saddle.highShores)
         self.assertEqual(newSaddle.edgeEffect, saddle.edgeEffect)
         self.assertEqual(newSaddle.edgePoints, saddle.edgePoints)
@@ -71,7 +71,7 @@ class SaddleTests(unittest.TestCase):
         a saddle with linked summits.
         """
         someslice = self.datamap.subset(0, 0, 30, 30)
-        domain = Domain(someslice)
+        domain = DomainMap(someslice)
         domain.run()
         saddles = domain.saddles
         saddle = saddles[7]
@@ -81,7 +81,7 @@ class SaddleTests(unittest.TestCase):
         self.assertEqual(newSaddle.latitude, saddle.latitude)
         self.assertEqual(newSaddle.longitude, saddle.longitude)
         self.assertEqual(newSaddle.elevation, saddle.elevation)
-        self.assertEqual(newSaddle.multiPoint, saddle.multiPoint)
+        self.assertEqual(newSaddle.multipoint, saddle.multipoint)
         self.assertEqual(newSaddle.highShores, saddle.highShores)
         self.assertEqual(newSaddle.edgeEffect, saddle.edgeEffect)
         self.assertEqual(newSaddle.edgePoints, saddle.edgePoints)
@@ -100,7 +100,7 @@ class SaddleTests(unittest.TestCase):
          saddle which has a parent.
         """
         someslice = self.datamap.subset(0, 0, 30, 30)
-        domain = Domain(someslice)
+        domain = DomainMap(someslice)
         domain.run()
         saddles = domain.saddles
         saddle = saddles[18]
@@ -111,7 +111,7 @@ class SaddleTests(unittest.TestCase):
         self.assertEqual(newSaddle.latitude, saddle.latitude)
         self.assertEqual(newSaddle.longitude, saddle.longitude)
         self.assertEqual(newSaddle.elevation, saddle.elevation)
-        self.assertEqual(newSaddle.multiPoint, saddle.multiPoint)
+        self.assertEqual(newSaddle.multipoint, saddle.multipoint)
         self.assertEqual(newSaddle.highShores, saddle.highShores)
         self.assertEqual(newSaddle.edgeEffect, saddle.edgeEffect)
         self.assertEqual(newSaddle.edgePoints, saddle.edgePoints)
@@ -273,7 +273,7 @@ class SaddleTests(unittest.TestCase):
         Single Point.
         """
         someslice = self.datamap.subset(0, 0, 30, 30)
-        domain = Domain(someslice)
+        domain = DomainMap(someslice)
         domain.run()
         saddles = domain.saddles
         saddle = saddles[7]
@@ -288,7 +288,7 @@ class SaddleTests(unittest.TestCase):
         Multipoint
         """
         someslice = self.datamap.subset(0, 0, 30, 30)
-        domain = Domain(someslice)
+        domain = DomainMap(someslice)
         domain.run()
         saddles = domain.saddles
         saddle = saddles.multipoints[1]